Add a configurable namespace to containers
When a user sets a default namespace on a container, that namespace will not be used in any container registrations for components from within that namespace. This makes it easier to reference them later.

For example, in a container with a namespace of “foo.bar”, a component at “foo/bar/baz.rb” will be accessible on the identifier “baz” only (with the default namespace removed for convenience).