Allow dry-auto_inject options to be passed to `Container#injector`
This makes it possible to provide the underlying injector with a custom strategies container (and any other config that may be made available in the future).

This requires some underlying changes, however: the `Inject` constant on `Container` subclasses is no longer set up automatically, since we can’t infer whether the user will want to pass their own options to the injector. Instead, the recommend practice should be that the user creates their own constant to store their preferred injector.