Stop auto-enabling :env plugin via :logging
We can't assue default :env plugin settings can be used, so it's better
to let the user enable :env manually. We should probably raise an error
when another required plugin is not enabled, but this needs a bit more
work so I'm gonna postpone it for now, but I'll get to it prior 1.0 for
sure.