Allow multiple lazy resolutions using same root as bootable component
There was a bug where a resolution for an object using the same root key as a bootable component (e.g. resolving “kitten_service.fetch_kitten” where there is a bootable component named :kitten_service) would only allow the _first_ object resolution to work. Subsequent resolutions (e.g. for “kitten_service.submit_kitten”) would raise Dry::Container::Error with a message like ‘Nothing registered with the key "kitten_service.submit_kitten”’ (even when this component did actually exist).

The problem exists only when using a non-finalized container (i.e. when the container is lazy loading components). On the first resolution, `Dry::System::Container.load_component` doesn’t detect that there is a bootable component on the root_key, and calls `.load_local_component`, which resolves the named component _and_ also properly detects and boots the bootable component named by the root key. On the subsequent resolution, `.load_component` _does_ detect that the bootable component exists, and just tries to boot it again, totally skipping the call to `.load_local_component`, which means the named component never gets required/loaded, and we get the error.

At the root (heh) of this, this really seems to be a problem of muddled responsibilities between `.load_component` and `.load_local_component` when it comes to booting bootable components, as well as two different ways to detect bootable components, only one of which works when the container is non-finalized and has not previously loaded or booted that component.

Fixing the root cause will take a little more time and thinking, and bigger adjustments.

In the meantime, this bug is effectively resolved by changing `.load_component` to always check to see if the named component key exists after booting and components or importing other containers. If the key doesn’t exist at that point, _always_ run `.load_local_component` to find the matching file, require it, and register the component. If that fails, then it can raise a (legitimate) “Nothing registered with the key” error.