Fix redefinition warnings happening on spec/integration
- spec/integration/external_components

The ExternalComponent::Loger class is defined in the :init block of
components/logger, and the system under test has two booters:

   :logger
   :my_logger, both

These booters are using the same external component, executing
the :init proc twice as expected on the lifecycle.

Since the definition is executed (twice) in a proc triggered inside
the RSpec example it escapes the removal done on the RSpec.before hook.

Using a guard avoid the redefinition, fixing the warnings.

- spec/integration/../dependency_graph_spec

The removal of ExternalComponent module was missing, this should
be done as we do on external_components_spec.rb