Introduce multiple bootable dirs
Add a bootable_dirs setting to Dry::System::Container and update the Booter so that it finds boot files from one or more directories.

For each configured bootable_dirs entry, relative directories will be appended to the configured container root, and absolute directories will be left unchanged.

When searching for bootable files, the first match will win, and any subsequent same-named files will not be loaded. In this way, the bootable_dirs acts similarly to the $PATH in a shell environment.

This arrangement will support uses where a container vendor may want to ship their own “default” bootable components, but allow the _user_ of the container to replace any of them one by one. To achieve this, the user can add their own entry to the beginning of the bootable_dirs setting, which will allow their own version of the bootable components to take precedence.