Make load_component easier to understand
Along with this, allow the standard Dry::Container::Error to be raised when an expected local component cannot be loaded. The ComponentLoadError wasn’t providing much in terms of additional useful information, and it was only raised in one particular circumstance. By removing this and allowing Dry::Container’s own error to bubble up, we end up with a more consistent set of behaviours around missing component errors.