Remove `use` method from provider source
This can be (and has been) confused for activating plugins, which use the `use` method on the container, and given we now have consistent and easy-to-understand access to the target container via `target`, it's more straightforward to use `target.start :some_provider` instead of a provider source-specifiic method like `#use`.

In addition, this is more flexible, since you can also access the other lifecycle steps for providers now too, like `target.prepare :some_provider`.