Make Container hooks more useful for dynamic registration (#274)
* Introduce after_register hook

Allows you to take action when a new key is registered, so that you can
decorate the key without knowing its name ahead of time.

See also hanami/hanami#1360

* Run after_finalize hooks before freeze

This hook is much less useful that it might be if you could still make
changes to the container when it ran.

* Add docs for before, after hooks