Use Zeitwerk::Registry to unload loaders in tests
Previously I enabled ObjectSpace in JRuby in zeitwerk_helpers.rb
programatically, but it did not fully work.
ObjectSpace.each_object(Zeitwerk::Loader) kept returning empty list. I
suppose this was because the ObjectSpace was enabled too late. It worked
correctly when JRUBY_OPTS="-X+O" wever passed in the command line.

However, because the CI config is set via repo automation, it was hard
to configure it just for dry-systems and would have a performance
penalty if we enabled it for all gems.

So instead, I figured out a way to drop usage of ObjectSpace in favour
of Zeitwer's Registry module, which does the same thing (presumably even
better).