Track and unregister loaders in tests manually (#293)
* Use Zeitwerk::Registry to unload loaders in tests

Previously I enabled ObjectSpace in JRuby in zeitwerk_helpers.rb
programatically, but it did not fully work.
ObjectSpace.each_object(Zeitwerk::Loader) kept returning empty list. I
suppose this was because the ObjectSpace was enabled too late. It worked
correctly when JRUBY_OPTS="-X+O" wever passed in the command line.

However, because the CI config is set via repo automation, it was hard
to configure it just for dry-systems and would have a performance
penalty if we enabled it for all gems.

So instead, I figured out a way to drop usage of ObjectSpace in favour
of Zeitwer's Registry module, which does the same thing (presumably even
better).

* Track and unregister loaders in tests manually

This replaces ZeitwerkHelpers with ZeitwerkLoaderRegistry, which keeps
track of created loaders and allows to unregister them. This is used in
tests when we need to have multiple loaders but cannot allow them to
interfere with each other.

This replaces former solutions to the problem:
- Using Zeitwerk::Loaders - it's a private API, not guaranteed to
  persist
- Using ObjectSpace - this does not work well in JRuby

* Use mocks for Zeitwerk::Loader.new

This ensures every call to .new in tests is captured and then
subsequently cleared via ZeitwerkLoaderRegistry.clear